explicitly set topology and disable CPU hotplug
Until now we omitted the topology definition in which case libvirt seems
to setup one socket per VCPU. The way multithreading works on Windows
this increases the number of context switches compared to when using
cores instead.

This disables CPU hotplug again!

A possible way out is by setting core to the maximum of assignable CPUs and set
the current attribute in the vcpu element to the desired number of CPUs
available in the VM at that time. Qemu should then fill up the defined
topology up to the number of current CPUs (leaving some CPUs unassigned).