ldaprecord: proper array handling in modify case
With overwrite set to false, current behavior is to create nested arrays
if the value to assign to an array attribute is an array itself.

Since the schema does not parse the attribute definitions (completely),
we may have NULL as a value for the existing attribute (instead of an
empty array). Make sure that we initialize it in this case.

This fixes #75